requirements.txt
Flask
boto3
python-dotenv
pandas
openpyxl
gunicorn
requests
